+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.2.0
+##### Minecraft 1.21.1
+
+1) Send world border only when it needed
+2) Fix few errors with world border calculation
+3) Add suggestion for luckperms
+4) Change from `radius` to `distance` in luckperms context
+
 ### 0.1.0
 ##### Minecraft 1.21.1